rewrote delivery algorithm, simplified producer flow control, various other optimisations, fixes and shit like that

+++ src/main/org/hornetq/api/core/management/QueueControl.java
@@ -80,7 +80,7 @@ public interface QueueControl
    /**
     * Returns the number of messages added to this queue since it was created.
     */
-   int getMessagesAdded();
+   long getMessagesAdded();
 
    /**
     * Returns the expiry address associated to this queue.
